acpi: BGRT table
Added support for the BGRT table (contains a pointer to the bootsplash bmp the firmware
displayed). It is used as the background for the terminal if present.
