+++ firefly/kernel/console/term.c
@@ -1,23 +1,16 @@
-#include "tterm.h"
-#include "gterm.h"
-#include "term.h"
+#include "firefly/console/gterm.h"
+#include "firefly/console/term.h"
 
-void term_init(struct term_t *term, callback_t callback, bool bios)
+void term_init(struct term_t *term, callback_t callback)
 {
     if (term->initialised == true)
         return;
 
     term->callback = callback;
-    term->bios = bios;
     term->term_backend = NOT_READY;
     term->arg = (uint64_t)term;
 
     term->gterm = alloc_mem(sizeof(struct gterm_t));
-#if defined(__i386__) || defined(__x86_64__)
-    if (bios == true)
-        term->tterm = alloc_mem(sizeof(struct tterm_t));
-#endif
-
     term->initialised = true;
 }
 
@@ -67,31 +60,12 @@ void term_vbe(struct term_t *term, struct framebuffer_t frm, struct font_t font,
     if (term->term_backend != VBE)
         term_deinit(term);
 
-    if (!gterm_init(term->gterm, term, frm, font, style, back) && term->bios)
-    {
-#if defined(__i386__) || defined(__x86_64__)
-        term_textmode(term);
-#endif
+    if (!gterm_init(term->gterm, term, frm, font, style, back))
         return;
-    }
 
     term_reinit(term);
     term->term_backend = VBE;
 }
-
-#if defined(__i386__) || defined(__x86_64__)
-void term_textmode(struct term_t *term)
-{
-    if (term->initialised == false || !term->tterm)
-        return;
-
-    term_deinit(term);
-    tterm_init(term->tterm, term);
-    term_reinit(term);
-
-    term->term_backend = TEXTMODE;
-}
-#endif
 
 void term_notready(struct term_t *term)
 {
@@ -965,10 +939,6 @@ void term_raw_putchar(struct term_t *term, uint8_t c)
 
     if (term->term_backend == VBE && term->gterm)
         gterm_putchar(term->gterm, c);
-#if defined(__i386__) || defined(__x86_64__)
-    else if (term->term_backend == TEXTMODE && term->tterm)
-        tterm_putchar(term->tterm, c);
-#endif
 }
 void term_clear(struct term_t *term, bool move)
 {
@@ -977,10 +947,6 @@ void term_clear(struct term_t *term, bool move)
 
     if (term->term_backend == VBE && term->gterm)
         gterm_clear(term->gterm, move);
-#if defined(__i386__) || defined(__x86_64__)
-    else if (term->term_backend == TEXTMODE && term->tterm)
-        tterm_clear(term->tterm, move);
-#endif
 }
 void term_enable_cursor(struct term_t *term)
 {
@@ -989,10 +955,6 @@ void term_enable_cursor(struct term_t *term)
 
     if (term->term_backend == VBE && term->gterm)
         gterm_enable_cursor(term->gterm);
-#if defined(__i386__) || defined(__x86_64__)
-    else if (term->term_backend == TEXTMODE && term->tterm)
-        tterm_enable_cursor(term->tterm);
-#endif
 }
 bool term_disable_cursor(struct term_t *term)
 {
@@ -1001,10 +963,7 @@ bool term_disable_cursor(struct term_t *term)
 
     if (term->term_backend == VBE && term->gterm)
         return gterm_disable_cursor(term->gterm);
-#if defined(__i386__) || defined(__x86_64__)
-    else if (term->term_backend == TEXTMODE && term->tterm)
-        return tterm_disable_cursor(term->tterm);
-#endif
+
     return false;
 }
 void term_set_cursor_pos(struct term_t *term, size_t x, size_t y)
@@ -1014,10 +973,6 @@ void term_set_cursor_pos(struct term_t *term, size_t x, size_t y)
 
     if (term->term_backend == VBE && term->gterm)
         gterm_set_cursor_pos(term->gterm, x, y);
-#if defined(__i386__) || defined(__x86_64__)
-    else if (term->term_backend == TEXTMODE && term->tterm)
-        tterm_set_cursor_pos(term->tterm, x, y);
-#endif
 }
 void term_get_cursor_pos(struct term_t *term, size_t *x, size_t *y)
 {
@@ -1026,10 +981,6 @@ void term_get_cursor_pos(struct term_t *term, size_t *x, size_t *y)
 
     if (term->term_backend == VBE && term->gterm)
         gterm_get_cursor_pos(term->gterm, x, y);
-#if defined(__i386__) || defined(__x86_64__)
-    else if (term->term_backend == TEXTMODE && term->tterm)
-        tterm_get_cursor_pos(term->tterm, x, y);
-#endif
 }
 void term_set_text_fg(struct term_t *term, size_t fg)
 {
@@ -1038,10 +989,6 @@ void term_set_text_fg(struct term_t *term, size_t fg)
 
     if (term->term_backend == VBE && term->gterm)
         gterm_set_text_fg(term->gterm, fg);
-#if defined(__i386__) || defined(__x86_64__)
-    else if (term->term_backend == TEXTMODE && term->tterm)
-        tterm_set_text_fg(term->tterm, fg);
-#endif
 }
 void term_set_text_bg(struct term_t *term, size_t bg)
 {
@@ -1050,10 +997,6 @@ void term_set_text_bg(struct term_t *term, size_t bg)
 
     if (term->term_backend == VBE && term->gterm)
         gterm_set_text_bg(term->gterm, bg);
-#if defined(__i386__) || defined(__x86_64__)
-    else if (term->term_backend == TEXTMODE && term->tterm)
-        tterm_set_text_bg(term->tterm, bg);
-#endif
 }
 void term_set_text_fg_bright(struct term_t *term, size_t fg)
 {
@@ -1062,10 +1005,6 @@ void term_set_text_fg_bright(struct term_t *term, size_t fg)
 
     if (term->term_backend == VBE && term->gterm)
         gterm_set_text_fg_bright(term->gterm, fg);
-#if defined(__i386__) || defined(__x86_64__)
-    else if (term->term_backend == TEXTMODE && term->tterm)
-        tterm_set_text_fg_bright(term->tterm, fg);
-#endif
 }
 void term_set_text_bg_bright(struct term_t *term, size_t bg)
 {
@@ -1074,10 +1013,6 @@ void term_set_text_bg_bright(struct term_t *term, size_t bg)
 
     if (term->term_backend == VBE && term->gterm)
         gterm_set_text_bg_bright(term->gterm, bg);
-#if defined(__i386__) || defined(__x86_64__)
-    else if (term->term_backend == TEXTMODE && term->tterm)
-        tterm_set_text_bg_bright(term->tterm, bg);
-#endif
 }
 void term_set_text_fg_default(struct term_t *term)
 {
@@ -1086,10 +1021,6 @@ void term_set_text_fg_default(struct term_t *term)
 
     if (term->term_backend == VBE && term->gterm)
         gterm_set_text_fg_default(term->gterm);
-#if defined(__i386__) || defined(__x86_64__)
-    else if (term->term_backend == TEXTMODE && term->tterm)
-        tterm_set_text_fg_default(term->tterm);
-#endif
 }
 void term_set_text_bg_default(struct term_t *term)
 {
@@ -1098,10 +1029,6 @@ void term_set_text_bg_default(struct term_t *term)
 
     if (term->term_backend == VBE && term->gterm)
         gterm_set_text_bg_default(term->gterm);
-#if defined(__i386__) || defined(__x86_64__)
-    else if (term->term_backend == TEXTMODE && term->tterm)
-        tterm_set_text_bg_default(term->tterm);
-#endif
 }
 bool term_scroll_disable(struct term_t *term)
 {
@@ -1110,10 +1037,7 @@ bool term_scroll_disable(struct term_t *term)
 
     if (term->term_backend == VBE && term->gterm)
         return gterm_scroll_disable(term->gterm);
-#if defined(__i386__) || defined(__x86_64__)
-    else if (term->term_backend == TEXTMODE && term->tterm)
-        return tterm_scroll_disable(term->tterm);
-#endif
+
     return false;
 }
 void term_scroll_enable(struct term_t *term)
@@ -1123,10 +1047,6 @@ void term_scroll_enable(struct term_t *term)
 
     if (term->term_backend == VBE && term->gterm)
         gterm_scroll_enable(term->gterm);
-#if defined(__i386__) || defined(__x86_64__)
-    else if (term->term_backend == TEXTMODE && term->tterm)
-        tterm_scroll_enable(term->tterm);
-#endif
 }
 void term_move_character(struct term_t *term, size_t new_x, size_t new_y, size_t old_x, size_t old_y)
 {
@@ -1135,10 +1055,6 @@ void term_move_character(struct term_t *term, size_t new_x, size_t new_y, size_t
 
     if (term->term_backend == VBE && term->gterm)
         gterm_move_character(term->gterm, new_x, new_y, old_x, old_y);
-#if defined(__i386__) || defined(__x86_64__)
-    else if (term->term_backend == TEXTMODE && term->tterm)
-        tterm_move_character(term->tterm, new_x, new_y, old_x, old_y);
-#endif
 }
 void term_scroll(struct term_t *term)
 {
@@ -1147,10 +1063,6 @@ void term_scroll(struct term_t *term)
 
     if (term->term_backend == VBE && term->gterm)
         gterm_scroll(term->gterm);
-#if defined(__i386__) || defined(__x86_64__)
-    else if (term->term_backend == TEXTMODE && term->tterm)
-        tterm_scroll(term->tterm);
-#endif
 }
 void term_revscroll(struct term_t *term)
 {
@@ -1159,10 +1071,6 @@ void term_revscroll(struct term_t *term)
 
     if (term->term_backend == VBE && term->gterm)
         gterm_revscroll(term->gterm);
-#if defined(__i386__) || defined(__x86_64__)
-    else if (term->term_backend == TEXTMODE && term->tterm)
-        tterm_revscroll(term->tterm);
-#endif
 }
 void term_swap_palette(struct term_t *term)
 {
@@ -1171,10 +1079,6 @@ void term_swap_palette(struct term_t *term)
 
     if (term->term_backend == VBE && term->gterm)
         gterm_swap_palette(term->gterm);
-#if defined(__i386__) || defined(__x86_64__)
-    else if (term->term_backend == TEXTMODE && term->tterm)
-        tterm_swap_palette(term->tterm);
-#endif
 }
 void term_save_state(struct term_t *term)
 {
@@ -1183,10 +1087,6 @@ void term_save_state(struct term_t *term)
 
     if (term->term_backend == VBE && term->gterm)
         gterm_save_state(term->gterm);
-#if defined(__i386__) || defined(__x86_64__)
-    else if (term->term_backend == TEXTMODE && term->tterm)
-        tterm_save_state(term->tterm);
-#endif
 
     term->context.saved_state_bold = term->context.bold;
     term->context.saved_state_reverse_video = term->context.reverse_video;
@@ -1205,10 +1105,6 @@ void term_restore_state(struct term_t *term)
 
     if (term->term_backend == VBE && term->gterm)
         gterm_restore_state(term->gterm);
-#if defined(__i386__) || defined(__x86_64__)
-    else if (term->term_backend == TEXTMODE && term->tterm)
-        tterm_restore_state(term->tterm);
-#endif
 }
 
 void term_double_buffer_flush(struct term_t *term)
@@ -1218,10 +1114,6 @@ void term_double_buffer_flush(struct term_t *term)
 
     if (term->term_backend == VBE && term->gterm)
         gterm_double_buffer_flush(term->gterm);
-#if defined(__i386__) || defined(__x86_64__)
-    else if (term->term_backend == TEXTMODE && term->tterm)
-        tterm_double_buffer_flush(term->tterm);
-#endif
 }
 
 uint64_t term_context_size(struct term_t *term)
@@ -1233,10 +1125,6 @@ uint64_t term_context_size(struct term_t *term)
 
     if (term->term_backend == VBE && term->gterm)
         ret += gterm_context_size(term->gterm);
-#if defined(__i386__) || defined(__x86_64__)
-    else if (term->term_backend == TEXTMODE && term->tterm)
-        ret += tterm_context_size(term->tterm);
-#endif
 
     return ret;
 }
@@ -1250,10 +1138,6 @@ void term_context_save(struct term_t *term, uint64_t ptr)
 
     if (term->term_backend == VBE && term->gterm)
         gterm_context_save(term->gterm, ptr);
-#if defined(__i386__) || defined(__x86_64__)
-    else if (term->term_backend == TEXTMODE && term->tterm)
-        tterm_context_save(term->tterm, ptr);
-#endif
 }
 void term_context_restore(struct term_t *term, uint64_t ptr)
 {
@@ -1265,10 +1149,6 @@ void term_context_restore(struct term_t *term, uint64_t ptr)
 
     if (term->term_backend == VBE && term->gterm)
         gterm_context_restore(term->gterm, ptr);
-#if defined(__i386__) || defined(__x86_64__)
-    else if (term->term_backend == TEXTMODE && term->tterm)
-        tterm_context_restore(term->tterm, ptr);
-#endif
 }
 void term_full_refresh(struct term_t *term)
 {
@@ -1276,8 +1156,4 @@ void term_full_refresh(struct term_t *term)
 
     if (term->term_backend == VBE && term->gterm)
         gterm_full_refresh(term->gterm);
-#if defined(__i386__) || defined(__x86_64__)
-    else if (term->term_backend == TEXTMODE && term->tterm)
-        tterm_full_refresh(term->tterm);
-#endif
 }
