+++ firefly/kernel/graphics/framebuffer.cpp
@@ -1,0 +1,5 @@
+#include "firefly/graphics/framebuffer.hpp"
+
+namespace firefly::kernel {
+FramebufferDevice fbDev;
+}  // namespace firefly::kernel
