+++ firefly/kernel/intel64/int/interrupt.cpp
@@ -86,7 +86,6 @@ void init() {
 void interrupt_handler(iframe iframe) {
     info_logger << "Int#: " << iframe.int_no << "\nError code: " << iframe.err << logger::endl;
     info_logger << "RIP: " << info_logger.hex(iframe.rip) << logger::endl;
-    backtrace(iframe.rip);
 
     for (;;)
         asm("cli\nhlt");
