+++ include/firefly/intel64/acpi/tables/all.hpp
@@ -1,0 +1,4 @@
+#pragma once
+
+#include "bgrt.hpp"
+#include "madt.hpp"
