+++ include/firefly/intel64/acpi/tables/madt.hpp
@@ -1,5 +1,1 @@
 #pragma once
-
-namespace firefly::kernel::core::acpi {
-
-}
