+++ include/firefly/panic.hpp
@@ -5,17 +5,16 @@
 
 namespace firefly {
 
-[[gnu::used]]
-[[noreturn]] static void panic(const char *msg) {
+[[gnu::used]] [[noreturn]] static void
+panic(const char *msg) {
     kernel::info_logger << "\n**** Kernel panic ****\nReason: " << msg << "\n";
-    // trace::trace_callstack();
+    trace::trace_callstack();
 
     while (1)
         asm volatile("hlt");
 }
 
-[[gnu::used]]
-[[noreturn]] static void
+[[gnu::used]] [[noreturn]] static void
 assertion_failure_panic(const char *msg) {
     kernel::info_logger << "\n**** Kernel panic ****\nAssertion failed: `" << msg << "`\n";
     trace::trace_callstack();
